+ BT parameters + medianFilter + MFA + Mileage

+++ Arduino/Digifiz/speedometer.h
@@ -2,10 +2,11 @@
 #define SPEEDOMETER_H
 #include "Arduino.h"
 #include <PinChangeInterrupt.h>
+#include <MedianFilterLib2.h>
 
 #define SPD_M_PIN PJ3
 
 void initSpeedometer();
-uint16_t readLastSpeed();
+uint32_t readLastSpeed();
 
 #endif
